+++ package.json
@@ -56,6 +56,7 @@
     "@jsii/spec": "^1.30.0",
     "case": "^1.6.3",
     "fs-extra": "^9.1.0",
+    "glob": "^7.1.7",
     "jsii-reflect": "^1.30.0",
     "yargs": "^16.2.0"
   },
