+++ src/cli.ts
@@ -1,22 +1,48 @@
+import * as fs from 'fs';
+import * as path from 'path';
+import * as glob from 'glob';
+import * as reflect from 'jsii-reflect';
 import * as yargs from 'yargs';
-import { renderFiles, renderSinglePageModule } from './index';
+import { Documentation } from './index';
 
 export async function main() {
   const args = yargs
-    .usage('Usage: $0 [JSII-FILE,...]')
-    .option('output', { type: 'string', alias: 'o', required: false, desc: 'Output directory or file (default directory is "dist", default file is "APP.md")' })
+    .usage('Usage: $0')
+    .option('output', { type: 'string', alias: 'o', required: false, desc: 'Output filename (defaults to API.md)' })
     .example('$0', 'Generate documentation for the current module as a single file (auto-resolves node depedencies)')
-    .example('$0 node_modules/**/.jsii', 'Generate documentation for all jsii modules in your project')
     .argv;
 
+  const assembly = await createAssembly();
 
-  if (args._.length === 0) {
-    const output = args.output ?? 'API.md';
-    await renderSinglePageModule(process.cwd(), output);
-  } else {
-    const output = args.output ?? 'dist';
-    await renderFiles(args._ as string[], output);
+  const docs = new Documentation({
+    assembly,
+    language: 'ts',
+    readme: false,
+  });
+
+  const output = args.output ?? 'API.md';
+
+  const markdown = docs.render().render();
+  fs.writeFileSync(output, markdown);
+}
+
+async function createAssembly(): Promise<reflect.Assembly> {
+
+  const ts = new reflect.TypeSystem();
+  const root = path.join(process.cwd(), '.jsii');
+
+  if (!fs.existsSync(root)) {
+    throw new Error(`Unable to locate jsii assembly at: ${root}. Make sure to run this command `
+      + 'from the root directory of your package after the jsii assembly has been created.');
   }
+
+  await ts.loadFile('.jsii', { isRoot: true });
+  for (let dependency of glob.sync('./node_modules/**/.jsii')) {
+    await ts.load(dependency);
+  }
+
+  return ts.roots[0];
+
 }
 
 main().catch(e => {
