+++ test/docgen/view/class.test.ts
@@ -1,0 +1,23 @@
+import * as reflect from 'jsii-reflect';
+import { PythonTranspile } from '../../../src/docgen/transpile/python';
+import { TypeScriptTranspile } from '../../../src/docgen/transpile/typescript';
+import { Class } from '../../../src/docgen/view/class';
+import { Assemblies } from '../assemblies';
+
+const assembly: reflect.Assembly = Assemblies.instance.withoutSubmodules;
+
+describe('python', () => {
+  const transpile = new PythonTranspile();
+  test('snapshot', () => {
+    const klass = new Class(transpile, assembly.classes[0]);
+    expect(klass.render().render()).toMatchSnapshot();
+  });
+});
+
+describe('typescript', () => {
+  const transpile = new TypeScriptTranspile();
+  test('snapshot', () => {
+    const klass = new Class(transpile, assembly.classes[0]);
+    expect(klass.render().render()).toMatchSnapshot();
+  });
+});
