+++ test/docgen/view/documentation.test.ts
@@ -1,0 +1,48 @@
+import * as reflect from 'jsii-reflect';
+import { Documentation } from '../../../src';
+import { Assemblies } from '../assemblies';
+
+const assembly: reflect.Assembly = Assemblies.instance.withoutSubmodules;
+const assemblyWithSubmodules: reflect.Assembly = Assemblies.instance.withSubmodules;
+
+describe('python', () => {
+  test('snapshot - root module', () => {
+    const docs = new Documentation({
+      language: 'python',
+      assembly: assembly,
+      readme: true,
+    });
+    expect(docs.render().render()).toMatchSnapshot();
+  });
+
+  test('snapshot - submodules', () => {
+    const docs = new Documentation({
+      language: 'python',
+      assembly: assemblyWithSubmodules,
+      submoduleName: 'aws_eks',
+      readme: true,
+    });
+    expect(docs.render().render()).toMatchSnapshot();
+  });
+});
+
+describe('typescript', () => {
+  test('snapshot - single module', () => {
+    const docs = new Documentation({
+      language: 'ts',
+      assembly: assembly,
+      readme: true,
+    });
+    expect(docs.render().render()).toMatchSnapshot();
+  });
+
+  test('snapshot - submodules', () => {
+    const docs = new Documentation({
+      language: 'ts',
+      assembly: assemblyWithSubmodules,
+      submoduleName: 'aws_eks',
+      readme: true,
+    });
+    expect(docs.render().render()).toMatchSnapshot();
+  });
+});
